dev_requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
boto3==1.9.221
botocore==1.12.221
docutils==0.15.2
entrypoints==0.3
flake8==3.7.8
importlib-metadata==0.20
jmespath==0.9.4
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pytest==5.0.1
pytest-mock==1.10.4
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
thrift==0.11.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.6.0
